pyproject.toml
[tool.poetry]
name = "ml-com-modelo-shap"
version = "0.1.0"
description = ""
authors = ["Douglas Saturnino <douglasaturnino@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.12"
pandas = "^2.2.1"
pyyaml = "^6.0.1"
matplotlib = "^3.8.3"
seaborn = "^0.13.2"
scikit-learn = "^1.3.2"
feature-engine = "^1.6.2"
xgboost = "^2.0.3"
shap = "^0.44.1"
yellowbrick = "^1.5"
jupyter = "^1.0.0"
scikit-plot = "^0.3.7"
fastapi = "^0.110.1"
uvicorn = "^0.29.0"
streamlit = "^1.33.0"

[tool.ruff]
# Exclude a variety of commonly ignored directories.
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".ipynb_checkpoints",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pyenv",
    ".pytest_cache",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    ".vscode",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "site-packages",
    "venv",
]

# Same as Black.
line-length = 79
indent-width = 4

[tool.ruff.lint]
fixable = ["I"]


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
